Move producer bits to producer module.
